fix(parser): handle flags, prefixes, and suffixes in include files
In the toolchain, the parser holds the flags, prefixes and suffixes
and the assembler uses them when finalizing the assembly. The parsers
of include files, however, aren't known to the assembler. In addition,
the flags, prefixes, and suffixes in include files must be scoped to the
include file.

This commit adds a new function to post process include file parsing.
The included content is enclosed in a new assemble block and flags,
prefixes, and suffixes are perpended / appended to the content.