fix: restrict use of evasion modifiers to end of line
Until now, `~` and `@` would be replaced by the evasion suffix
regardless of their position in the string, even though the evasion
suffix was only ever intended to be used as a suffix. With this commit,
`~` and `@` are only treated specially if they occur as the last
character on the line.

In addition, it is now possible to escape a terminal `~` or `@` with a
backslash. The only limitation is that the toolchain cannot handle a
sequence such as `abc\\~`, which would result in `abc\\<suffix>` instead
of `abc\<suffix>` because there's no special handling of backslash
espaces in general.

Fixes #185