fix: meta character escaping
The cmdline processor treats each line of input as a string, not a
regular rexpression. However, the assembler treats input as regular
expressions. The output of the cmdline processor must, therefore, be
properly escaped.

The function responsible for handling escape sequences was missing the
`+` character, causing literal `+` characters to end up being passed to
the assembler.